// src/models/contact.model.ts
export class ContactModel {
  constructor() {
    this.firstName = "";
    this.lastName = "";
    this.email = "";
    this.phoneNumber = "";
    this.address = "";
    this.city = "";
    this.country = "";
    this.state = "";
    this.postalCode = "";
  }
  firstName: string;
  lastName: string;
  email: string;
  phoneNumber: string;
  address: string;
  city: string;
  state: string;
  country: string;
  postalCode: string;
}